requirements.txt
django == 2.2
pillow
django-filter